requirements.txt
pandas
scikit-learn
streamlit
arxiv
google-generativeai
openpyxl
joblib
nltk
python-dotenv
